+++ src/modules/delete-task.js
@@ -1,8 +1,11 @@
-function deleteTask(todoTasks, index) {
+import { iterateToDoTasks } from './iterate-tasks';
+
+export function deleteTask(todoTasks, index) {
   todoTasks.splice(index, 1);
   todoTasks.forEach((todoTasks, index) => {
     todoTasks.index = index;
+    iterateToDoTasks(); 
   });
 }
 
-export default deleteTask;
+let todoTasks = [];
